fl-health/requirements.txt
# ===== Core ML / FL =====
torch>=2.2
torchvision>=0.17
scikit-learn>=1.4
numpy>=1.26
pandas>=2.2
flwr>=1.8.0
opacus>=1.4   # DP

# ===== Utils =====
loguru>=0.7
pyyaml>=6.0.1
python-dotenv>=1.0.1

# ===== Blockchain (Ethereum-compatible) =====
web3>=6.11.4
eth-account>=0.11.0

# سنضيف HE (مثل SEAL/TenSEAL) لاحقًا في خطوة مخصّصة
